requirements.txt
cassandra_driver==3.25.0
Flask==1.1.2
wget==3.2
